Added discount

+++ packages/purchases/lib/core/models/purchases_state.dart
@@ -9,7 +9,7 @@ class PurchasesState with _$PurchasesState {
     String? threeMonthPrice,
     String? sixMonthPrice,
     String? annualPrice,
-    double? annualDiscount,
+    int? annualDiscount,
     @Default(false) bool isReady,
     @Default(false) bool hasPackage,
     @Default(false) bool isPurchasing,
